VCST-3529: Remove limitations to max facet value to 250.
fix: Added documentation for default facet value behavior in Elastic App Search, specifying the default of 10 and a maximum of 250. Updated logic for the Size property in ValueFacet to default to MaxFacetValues when termAggregationRequest.Size is 0, ensuring clarity and adherence to the maximum limit. You can change the number of facet values returned by app_search.engine.total_facet_values_returned.limit and then pass it using size value in TermAggregationRequest.